+++ GHEtool/VariableClasses/PipeData/__init__.py
@@ -3,3 +3,4 @@ from .SingleUTube import SingleUTube
 from .DoubleUTube import DoubleUTube
 from .MultipleUTube import MultipleUTube
 from .CoaxialPipe import CoaxialPipe
+from .Separatus import Separatus