+++ GHEtool/test/unit-tests/test_loads/test_hourly_load_building.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 
 from GHEtool import FOLDER
-from GHEtool.VariableClasses import HourlyBuildingLoad
+from GHEtool.VariableClasses import HourlyBuildingLoad, Cluster
 from GHEtool.VariableClasses.Result import ResultsMonthly, ResultsHourly
 
 from GHEtool.VariableClasses.Efficiency import *
@@ -645,3 +645,49 @@ def test_time_array():
     load.start_month = 2
     assert np.allclose(load.month_indices, np.tile(np.concatenate((
         np.repeat(np.arange(1, 13), load.UPM)[730:], np.repeat(np.arange(1, 13), load.UPM)[:730])), 10))
+
+
+def test_cluster():
+    load1 = HourlyBuildingLoad(np.linspace(1, 2000, 8760), np.linspace(1, 8760 - 1, 8760) * 2, 10, cop_basic, eer_basic)
+    load2 = HourlyBuildingLoad(np.linspace(1, 2000, 8760), np.linspace(1, 8760 - 1, 8760) * 2, 10, cop_basic, eer_basic)
+    load = HourlyBuildingLoad(np.linspace(1, 2000, 8760) * 2, np.linspace(1, 8760 - 1, 8760) * 2 * 2, 10, cop_basic,
+                              eer_basic)
+
+    cluster = Cluster([load1, load2])
+
+    assert np.allclose(load.monthly_baseload_extraction,
+                       cluster.monthly_baseload_extraction)
+    assert np.allclose(load.monthly_baseload_injection,
+                       cluster.monthly_baseload_injection)
+    assert np.allclose(load.monthly_peak_extraction,
+                       cluster.monthly_peak_extraction)
+    assert np.allclose(load.monthly_peak_injection,
+                       cluster.monthly_peak_injection)
+    assert np.allclose(load.hourly_extraction_load, cluster.hourly_extraction_load)
+    assert np.allclose(load.hourly_injection_load, cluster.hourly_injection_load)
+
+    load.reset_results(0, 10)
+    cluster.reset_results(0, 10)
+    assert np.allclose(load.hourly_extraction_load, cluster.hourly_extraction_load)
+    assert np.allclose(load.hourly_injection_load, cluster.hourly_injection_load)
+
+    load.set_results(results_hourly_test)
+    cluster.set_results(results_hourly_test)
+    assert np.allclose(load.hourly_extraction_load, cluster.hourly_extraction_load)
+    assert np.allclose(load.hourly_injection_load, cluster.hourly_injection_load)
+
+
+def test_repr_():
+    load = HourlyBuildingLoad()
+    load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
+    load.dhw = 10000
+
+    assert 'Hourly building load\n' \
+           'Efficiency heating: SCOP [-]: 5\n' \
+           'Efficiency cooling: SEER [-]: 20\n' \
+           'Peak cooling duration [hour]: 6.0\n' \
+           'Peak heating duration [hour]: 6.0\n' \
+           'Simulation period [year]: 20\n' \
+           'First month of simulation [-]: 1\n' \
+           'DHW demand [kWh/year]: 10000\n' \
+           'Efficiency DHW: SCOP [-]: 4' == load.__repr__()