+++ GHEtool/test/unit-tests/test_loads/test_monthly_load_building.py
@@ -633,3 +633,53 @@ def test_time_array():
     assert np.allclose(load.month_indices, np.tile(np.arange(1, 13), 10))
     load.start_month = 2
     assert np.allclose(load.month_indices, np.tile(np.concatenate((np.arange(2, 13), [1])), 10))
+
+
+def test_repr_():
+    load = MonthlyBuildingLoadAbsolute(*load_case(1), 10, 5, 5)
+    assert 'Monthly building load\n' \
+           'Month\tPeak heating [kW]\tPeak cooling [kW]\tBaseload heating [kWh]\t' \
+           'Baseload cooling [kWh]\n' \
+           '1\t63.70\t5.14\t46500.00\t3750.00\n' \
+           '2\t60.82\t10.27\t44400.00\t7500.00\n' \
+           '3\t51.37\t22.00\t37500.00\t7500.00\n' \
+           '4\t40.68\t44.00\t29700.00\t7500.00\n' \
+           '5\t26.30\t83.00\t19200.00\t11250.00\n' \
+           '6\t0.00\t117.00\t0.00\t15000.00\n' \
+           '7\t0.00\t134.00\t0.00\t30000.00\n' \
+           '8\t0.00\t150.00\t0.00\t30000.00\n' \
+           '9\t25.07\t100.00\t18300.00\t15000.00\n' \
+           '10\t35.75\t23.00\t26100.00\t11250.00\n' \
+           '11\t48.08\t10.27\t35100.00\t7500.00\n' \
+           '12\t59.18\t5.14\t43200.00\t3750.00\n' \
+           'Peak cooling duration [hour]: 6.0\n' \
+           'Peak heating duration [hour]: 6.0\n' \
+           'Efficiency heating: SCOP [-]: 5\n' \
+           'Efficiency cooling: SEER [-]: 5\n' \
+           'Simulation period [year]: 10\n' \
+           'First month of simulation [-]: 1' == load.__repr__()
+
+    load.dhw = 10000
+    assert 'Monthly building load\n' \
+           'Month\tPeak heating [kW]\tPeak cooling [kW]\tBaseload heating [kWh]\t' \
+           'Baseload cooling [kWh]\n' \
+           '1\t63.70\t5.14\t46500.00\t3750.00\n' \
+           '2\t60.82\t10.27\t44400.00\t7500.00\n' \
+           '3\t51.37\t22.00\t37500.00\t7500.00\n' \
+           '4\t40.68\t44.00\t29700.00\t7500.00\n' \
+           '5\t26.30\t83.00\t19200.00\t11250.00\n' \
+           '6\t0.00\t117.00\t0.00\t15000.00\n' \
+           '7\t0.00\t134.00\t0.00\t30000.00\n' \
+           '8\t0.00\t150.00\t0.00\t30000.00\n' \
+           '9\t25.07\t100.00\t18300.00\t15000.00\n' \
+           '10\t35.75\t23.00\t26100.00\t11250.00\n' \
+           '11\t48.08\t10.27\t35100.00\t7500.00\n' \
+           '12\t59.18\t5.14\t43200.00\t3750.00\n' \
+           'Peak cooling duration [hour]: 6.0\n' \
+           'Peak heating duration [hour]: 6.0\n' \
+           'Efficiency heating: SCOP [-]: 5\n' \
+           'Efficiency cooling: SEER [-]: 5\n' \
+           'Simulation period [year]: 10\n' \
+           'First month of simulation [-]: 1\n' \
+           'DHW demand [kWh/year]: 10000\n' \
+           'Efficiency DHW: SCOP [-]: 4' == load.__repr__()