+++ GHEtool/test/unit-tests/test_loads/test_monthly_load_geothermal.py
@@ -2,7 +2,7 @@ import pytest
 
 import numpy as np
 
-from GHEtool.VariableClasses import MonthlyGeothermalLoadAbsolute, HourlyGeothermalLoad, MonthlyGeothermalLoadMultiYear
+from GHEtool.VariableClasses import MonthlyGeothermalLoadAbsolute, HourlyGeothermalLoad, Cluster
 from GHEtool.Validation.cases import load_case
 
 
@@ -362,3 +362,54 @@ def test_yearly_loads():
 def test_depreciation_warning():
     with pytest.raises(DeprecationWarning):
         MonthlyGeothermalLoadAbsolute(baseload_heating=[1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12])
+
+
+def test_cluster():
+    load1 = MonthlyGeothermalLoadAbsolute(*load_case(1))
+    load2 = MonthlyGeothermalLoadAbsolute(*load_case(2))
+    cluster = Cluster([load1, load2])
+    cluster2 = Cluster()
+    cluster2.add_building(load1)
+    cluster2.add_building(load2)
+    assert np.allclose(load1.monthly_baseload_extraction + load2.monthly_baseload_extraction,
+                       cluster.monthly_baseload_extraction)
+    assert np.allclose(load1.monthly_baseload_injection + load2.monthly_baseload_injection,
+                       cluster.monthly_baseload_injection)
+    assert np.allclose(load1.monthly_peak_extraction + load2.monthly_peak_extraction,
+                       cluster.monthly_peak_extraction)
+    assert np.allclose(load1.monthly_peak_injection + load2.monthly_peak_injection,
+                       cluster.monthly_peak_injection)
+    assert np.allclose(load1.monthly_baseload_extraction + load2.monthly_baseload_extraction,
+                       cluster2.monthly_baseload_extraction)
+    assert np.allclose(load1.monthly_baseload_injection + load2.monthly_baseload_injection,
+                       cluster2.monthly_baseload_injection)
+    assert np.allclose(load1.monthly_peak_extraction + load2.monthly_peak_extraction,
+                       cluster2.monthly_peak_extraction)
+    assert np.allclose(load1.monthly_peak_injection + load2.monthly_peak_injection,
+                       cluster2.monthly_peak_injection)
+    cluster.simulation_period = 21
+    assert cluster.list_of_buildings[0].simulation_period == 21
+
+
+def test_repr_():
+    load = MonthlyGeothermalLoadAbsolute(*load_case(1))
+
+    assert 'Monthly geothermal load\n' \
+           'Month\tPeak extraction [kW]\tPeak injection [kW]\tBaseload extraction ' \
+           '[kWh]\tBaseload injection [kWh]\n' \
+           '1\t63.70\t5.14\t46500.00\t3750.00\n' \
+           '2\t60.82\t10.27\t44400.00\t7500.00\n' \
+           '3\t51.37\t22.00\t37500.00\t7500.00\n' \
+           '4\t40.68\t44.00\t29700.00\t7500.00\n' \
+           '5\t26.30\t83.00\t19200.00\t11250.00\n' \
+           '6\t0.00\t117.00\t0.00\t15000.00\n' \
+           '7\t0.00\t134.00\t0.00\t30000.00\n' \
+           '8\t0.00\t150.00\t0.00\t30000.00\n' \
+           '9\t25.07\t100.00\t18300.00\t15000.00\n' \
+           '10\t35.75\t23.00\t26100.00\t11250.00\n' \
+           '11\t48.08\t10.27\t35100.00\t7500.00\n' \
+           '12\t59.18\t5.14\t43200.00\t3750.00\n' \
+           'Peak injection duration [hour]: 6.0\n' \
+           'Peak extraction duration [hour]: 6.0\n' \
+           'Simulation period [year]: 20\n' \
+           'First month of simulation [-]: 1' == load.__repr__()